% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_annot.R
\name{box_annot}
\alias{box_annot}
\title{box_annot}
\usage{
box_annot(x)
}
\arguments{
\item{x}{values is the x aesthetic in ggplot2 call}
}
\description{
annotate boxplots with test, based on x(factor) position
}
\examples{
ggplot() + stat_summary(aes(...), geom="text", fun.data=box_annot, label=list)
}
