% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunnett_contrasts.R
\name{dunnett_contrasts}
\alias{dunnett_contrasts}
\title{dunnett_contrasts}
\usage{
dunnett_contrasts(x, ref.index, factor, interaction)
}
\arguments{
\item{x}{input if model fit ie lm, glm, glmer etc...}

\item{ref.index}{index of reference level for multiple comparisons. You can find this
using names(<fit>$coefficients) for lm or glm obj or names(fixef(<fit>)) for mixed models}

\item{factor}{predictor variable for comparisons, usually a group ie genotype}

\item{interaction}{optional term for interaction predictor if relevant.}
}
\description{
generate Dunnett-adjusted multiple comparisons
using lsmeans package
}
\examples{
contrasts<-dunnett_contrasts(fit, ref.index = 1, factor = "genotype", interaction = "food")
}
