% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_dauer.R
\name{format_dauer}
\alias{format_dauer}
\title{format_dauer}
\usage{
format_dauer(df, p.dauer)
}
\arguments{
\item{df}{csv or data frame (can be piped)}

\item{p.dauer}{include post an/or partial dauers. Arguments "exclude" (default) - not counted,
"non" - animals are considered non-dauer in cases where 100% dauers are observed,
"dauer" - animals are counted as dauers in cases where only partial dauers form (daf-16)}
}
\description{
formats dauer data for analysis requires variables 'strains' and 'foods'
}
\examples{
df \%>\% format_dauer(p.dauer = "exclude")


}
