% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStan_CIs.R
\name{getStan_CIs}
\alias{getStan_CIs}
\title{getStan_CIs}
\usage{
getStan_CIs(x, type, group, base, compare_gt)
}
\arguments{
\item{x}{Stan glmm model}

\item{type}{type of data, options are NA (linear model), "log" (Log10), "roam" (Poisson), "dauer" (logit)}

\item{group}{optional argument for group-level effects, i.e. "food" or "temp". Must match column name for condition}

\item{base}{optional argument for log models - base of log function}

\item{compare_gt}{optional argument for group-level effect - if true, genotype is compared at each level of condition, if blank,
then condition effects are returned.}
}
\description{
Function pulls confidence intervals from a StanGLMM model using genotype as a parameter/predictor
}
\examples{
getStan_CIs(stan.glmm, type = "log", group = "food", base = 10, compare_gt = TRUE)

}
