% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_layers.R
\name{ggplot_layers}
\alias{ggplot_layers}
\alias{add.scatter}
\title{ggplot_layers}
\usage{
add.scatter()
}
\arguments{
\item{width}{optional width for median lines}
}
\description{
adds layers to ggplots
}
\section{}{
NA
}

\examples{
p <- ggplot(aes(x=genotype, y=pct)
p + add.scatter
}
