% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MM_CI_trans_dunnett.R
\name{MM_CI_trans_dunnett}
\alias{MM_CI_trans_dunnett}
\title{MM_CI_trans_dunnett}
\usage{
MM_CI_trans_dunnett(y, predictor)
}
\arguments{
\item{y}{input if model fit ie lm, glm, glmer etc...}

\item{factor}{optional predictor variable for comparisons, defaults to "genotype"}
}
\description{
generate Dunnett-adjusted 95 pct confidence intervals for plotting on the
untransformed response scale for binomial or other transformed model data
using lsmeans package
}
\examples{
mixed <- MM_CI_trans_dunnett(fit, predictor = "genotype")
}
