% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MM_CI_trans_Tukey.R
\name{MM_CI_trans_Tukey}
\alias{MM_CI_trans_Tukey}
\title{MM_CI_trans_Tukey}
\usage{
MM_CI_trans_Tukey(y, predictor)
}
\arguments{
\item{y}{input if model fit ie lm, glm, glmer etc...}

\item{predictor}{optional predictor variable for comparisons, defaults to "genotype"}
}
\description{
generate Tukey-adjusted 95 pct confidence intervals for plotting on the
untransformed response scale for binomial or other transformed model data
using lsmeans package. For glmms no degrees of freedom are available, so these
are asymptotic confidence intervals.
}
\examples{
mixed <- MM_CI_trans_Tukey(fit, predictor = "genotype")
}
