% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CIs_dark.R
\name{plot_CIs_dark}
\alias{plot_CIs_dark}
\title{plot_CIs_dark}
\usage{
plot_CIs_dark(df, title, plot.contrasts, plot.contrasts.2, ypos, type, offset)
}
\arguments{
\item{df}{input dataset. Requires a "genotype" column. Response needs to be "pct" for
dauer plots, "n_entries" for roaming data, and "cell.norm" for gene expression microscopy
data. Also requires a "mixed" dataset, generated by either MM_CIs.R, Mm_CI_trans_Tukey.R or MM_CI_trans_dunnett.R}

\item{title}{title for the plot}

\item{plot.contrasts}{list of pairwise comparisons to plot. Usually generated by selecting a list 
of comparisons using tukey_contrasts or dunnett_contrasts functions which use the lsmeans
package, ie plot.contrasts = c("",contrasts$prange[1:2]) for 2
comparisons to a control.}

\item{plot.contrasts.2}{list of secondary comparisons (ie rescue lines). These will be colored red and offset in the
y direction indicated by offset.}

\item{ypos}{y position of p value for plot.contrasts on the plot response scale}

\item{type}{type of plot ("dauer", "roam", "GFP"). Defaults to mRNA plot if no input.}

\item{offset}{position to offset plot.contrast.2 on the y response scale.}
}
\description{
Function generates boxplots for dauer, expression or roaming data along with
accompanying NHST comparisons and 95 pct confidence intervals
}
\examples{
plot_CIs(df, "GFP plot", plot.contrasts = contrasts, ypos = 500, type = "GFP", offset = 100)
}
