% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cooksD.R
\name{plot_cooksD}
\alias{plot_cooksD}
\title{plot_cooksD}
\usage{
plot_cooksD(lin.mod, threshold, noplot)
}
\arguments{
\item{lin.mod}{model fit, lm, glm, lmm, glmm}

\item{threshold}{threshold for outlier flagging. Usually 4 (4*mean cooks D)}

\item{noplot}{optional to plot outlier observations, leave blank for no plot}
}
\value{
a list of observation indices based on the original data frame
}
\description{
Identify highly influential observations based on cooks distance. 
Common outlier threshold is 4*mean of cooksD
}
\examples{
outliers.cooksd <- plot_cooksD(fit, noplot = TRUE)
}
