% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dauer_noCIs.R
\name{plot_dauer_noCIs}
\alias{plot_dauer_noCIs}
\title{plot_dauer_noCIs}
\usage{
plot_dauer_noCIs(df, title, fillsig, contrasts, level)
}
\arguments{
\item{df}{input dataset. Requires a "genotype" column. Response needs to be "pct" for
dauer plots.}

\item{title}{title for the plot}

\item{fillsig}{optional argument to include box fill colored by significance level specified in level argument}

\item{contrasts}{pairwise contrast object output from tukey_contrasts() or dunnett_contrasts()}

\item{level}{alpha level for filling box plots. White boxes are cases when p > level.}
}
\description{
Function generates boxplots for dauer data. Optionally fills boxes based on alpha cutoff 
level from tukey_contrasts or dunnett_contrasts
}
\examples{
plot_dauer_noCIs(df, title = "dauer plot", plot.contrasts = plot.contrasts, ypos =  1.075)
}
