% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dauer_SEs.R
\name{plot_dauer_SEs}
\alias{plot_dauer_SEs}
\title{plot_dauer_SEs}
\usage{
plot_dauer_SEs(df, title, plot.contrasts, plot.contrasts.2, ypos)
}
\arguments{
\item{df}{input dataset. Requires a "genotype" column. Response needs to be "pct" for
dauer plots. Also requires a "mixed" dataset, generated by either MM_CIs.R, Mm_CI_trans_Tukey.R or MM_CI_trans_dunnett.R}

\item{title}{title for the plot}

\item{plot.contrasts}{list of pairwise comparisons to plot. Usually generated by selecting a list 
of comparisons using tukey_contrasts or dunnett_contrasts functions which use the lsmeans
package, ie plot.contrasts = c("",contrasts$prange[1:2]) for 2
comparisons to a control.}

\item{plot.contrasts.2}{list of secondary comparisons (ie rescue lines). These will be colored red and offset in the
y direction indicated by offset.}

\item{ypos}{y position of p value for plot.contrasts on the plot response scale}
}
\description{
Function generates boxplots for dauer data along with standard error generate from a glm fit.
}
\examples{
plot_dauer_SEs(df, title = "dauer plot", plot.contrasts = plot.contrasts, ypos =  1.075)
}
