% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outlier.resids.R
\name{plot_outlier.resids}
\alias{plot_outlier.resids}
\title{plot_outlier.resids}
\usage{
plot_outlier.resids(lin.mod, noplot)
}
\arguments{
\item{lin.mod}{model fit, lm, glm, lmm, glmm}

\item{noplot}{optional to plot outlier observations, leave blank for no plot}
}
\value{
a list of observation indices based on the original data frame
}
\description{
identify outliers based on standardized residuals using MASS package. outliers are flagged if absolute standardized
residual is greater than 3 and is > 1.5*IQR for standardized residuals.
}
\examples{
outliers.q <- plot_outlier.resids(fit, noplot = TRUE)
}
