% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regression_se.R
\name{plot_regression_se}
\alias{plot_regression_se}
\title{plot_regression_se}
\usage{
plot_regression_se(df, meanX, meanY, semX, semY)
}
\arguments{
\item{df}{input dataset. Requires the following columns: meanX (pheromone), meanY (27degree),
semX and semY. Also a lat_abs column for latitude of each strain.}
}
\description{
plot correlation between pheromone dauer formation and high temp dauer formation using SEM, with a 
heat map for isolation latitude.
}
\examples{
plot_regression_se(C327, "mean.C3", "mean.HT","SEM.C3", "SEM.HT")
}
