% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_regression_seone.R
\name{plot_regression_seone}
\alias{plot_regression_seone}
\title{plot_regression_seone}
\usage{
plot_regression_seone(df, meanX, meanY, semX)
}
\arguments{
\item{df}{input dataset. Requires the following columns: meanX (pheromone), meanY, semX, and abs_lat}
}
\description{
plot dauer formation (x) vs. absolute latitidue (y)
}
\examples{
plot_regression_seone(df, "mean.HT","SEM.HT")
}
