% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prange.R
\name{prange}
\alias{prange}
\title{prange}
\usage{
prange(x)
}
\arguments{
\item{x}{contrasts object, must have p.value column}
}
\value{
an lsmeans contrast list with 'prange' column
}
\description{
convert p values to estimates using a contrast object from tukey_ or dunnett_contrasts.R
}
\examples{
contrasts <- prange(dunnett_contrasts(fit, ...))
}
