% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dauer_stan.R
\name{run_dauer_stan}
\alias{run_dauer_stan}
\title{run_dauer_stan}
\usage{
run_dauer_stan(df, type, parameters)
}
\arguments{
\item{df}{input dataset. Requires a "genotype" column. See type for data column types}

\item{type}{"dauer" (default),"dauer-grouped". For dauer data, needs to have raw counts, with "dauer" column and "n" column.
For grouped data, must have group.id column - usually interaction(genotype,condition)}
}
\description{
Function runs a stan glmm for dauer data
}
\examples{
 df \%>\% run_dauer_stan(parameters)
}
