% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey_contrasts.R
\name{tukey_contrasts}
\alias{tukey_contrasts}
\title{tukey_contrasts}
\usage{
tukey_contrasts(x, factor)
}
\arguments{
\item{x}{model fit of type lm, lmer, glmer, glm etc...}

\item{factor}{primary predictor factor for comparisons ie genotype, temperature etc...}
}
\value{
an lsmeans contrast list
}
\description{
generate post-hoc pairwise Tukey comparisons based on a model fit
results are presented on the original response scale
}
\examples{
contrasts <- tukey_contrasts(glmm.fit, "genotype")
}
